server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-multipart
pydantic
bcrypt
PyJWT
google-generativeai
PyMuPDF
python-docx
numpy
email-validator
sentence-transformers
faiss-cpu
